pyproject.toml
[project]
name = "multimodal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py>=2.3.0",
    "datasets>=3.6.0",
    "evaluate>=0.4.4",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.7.0",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "transformers==4.53.0",
    "wandb>=0.20.1",
]

